Move Linux-only integration/network tests into Linux-only file
